backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
email-validator==2.2.0
python-multipart==0.0.20
SQLAlchemy==2.0.38
PyMySQL==1.1.1
cryptography==44.0.1
python-dotenv==1.0.1

# auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

# tests
pytest==8.3.4
httpx==0.28.1

